refactor(frontend): move test files to dedicated tests/ directory
